+++ hydra-cardano-api/src/Hydra/Cardano/Api.hs
@@ -53,6 +53,7 @@ import Cardano.Api as X hiding (
   TxBodyScriptData (..),
   TxExtraKeyWitnesses (..),
   TxFee (..),
+  TxIns,
   TxInsCollateral (..),
   TxInsReference (..),
   TxMetadataInEra (..),
@@ -137,12 +138,10 @@ import Hydra.Cardano.Api.Witness as Extras
 
 import qualified Cardano.Api
 import qualified Cardano.Api.Shelley
-import qualified Cardano.Ledger.Alonzo.Data as Ledger
-import qualified Cardano.Ledger.Alonzo.Scripts as Ledger
 import qualified Cardano.Ledger.Alonzo.TxWitness as Ledger
-import qualified Cardano.Ledger.Babbage.TxBody as Ledger
+import qualified Cardano.Ledger.Core as Ledger
 import qualified Cardano.Ledger.Keys as Ledger
-import qualified Cardano.Ledger.Shelley.Address.Bootstrap as Ledger
+import qualified Cardano.Ledger.Keys.Bootstrap as Ledger
 import qualified Cardano.Ledger.Shelley.Tx as Ledger hiding (TxBody)
 import Data.ByteString.Short (ShortByteString)
 import Prelude
@@ -190,9 +189,9 @@ pattern ShelleyAddressInAnyEra <-
 type BalancedTxBody = Cardano.Api.BalancedTxBody Era
 {-# COMPLETE BalancedTxBody #-}
 
-pattern BalancedTxBody :: TxBody -> TxOut CtxTx -> Lovelace -> BalancedTxBody
-pattern BalancedTxBody{balancedTxBody, balancedTxChangeOutput, balancedTxFee} <-
-  Cardano.Api.BalancedTxBody balancedTxBody balancedTxChangeOutput balancedTxFee
+pattern BalancedTxBody :: TxBodyContent BuildTx -> TxBody -> TxOut CtxTx -> Lovelace -> BalancedTxBody
+pattern BalancedTxBody{balancedTxBodyContent, balancedTxBody, balancedTxChangeOutput, balancedTxFee} <-
+  Cardano.Api.BalancedTxBody balancedTxBodyContent balancedTxBody balancedTxChangeOutput balancedTxFee
   where
     BalancedTxBody =
       Cardano.Api.BalancedTxBody
@@ -615,14 +614,14 @@ pattern TxOutDatumHash{txOutDatumHash} <-
     TxOutDatumHash =
       Cardano.Api.TxOutDatumHash ScriptDataInBabbageEra
 
-pattern TxOutDatumInTx :: ScriptData -> TxOutDatum CtxTx
+pattern TxOutDatumInTx :: HashableScriptData -> TxOutDatum CtxTx
 pattern TxOutDatumInTx{txOutDatumScriptData} <-
   Cardano.Api.TxOutDatumInTx _ txOutDatumScriptData
   where
     TxOutDatumInTx =
       Cardano.Api.TxOutDatumInTx ScriptDataInBabbageEra
 
-pattern TxOutDatumInline :: ScriptData -> TxOutDatum ctx
+pattern TxOutDatumInline :: HashableScriptData -> TxOutDatum ctx
 pattern TxOutDatumInline{txOutDatumInlineScriptData} <-
   Cardano.Api.TxOutDatumInline _ txOutDatumInlineScriptData
   where